requirements.txt
python-telegram-bot==21.7
google-generativeai==0.7.2
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
python-dateutil==2.8.2
requests==2.31.0
Pillow>=10.0.0
httpx==0.27.2